Move zeroconf exit handling to server.py
