+++ kolibri/utils/cli.py
@@ -380,14 +380,6 @@ def update(old_version, new_version):
     cache.clear()
 
 
-def _cleanup_before_quitting(signum, frame):
-    from kolibri.core.discovery.utils.network.search import unregister_zeroconf_service
-
-    unregister_zeroconf_service()
-    signal.signal(signum, signal.SIG_DFL)
-    os.kill(os.getpid(), signum)
-
-
 main_help = """Kolibri management commands
 
 For more information, see:
@@ -410,13 +402,7 @@ def main(ctx):
     if ctx.invoked_subcommand is None:
         click.echo(ctx.get_help())
         ctx.exit(1)
-
-    try:
-        signal.signal(signal.SIGINT, _cleanup_before_quitting)
-        signal.signal(signal.SIGTERM, _cleanup_before_quitting)
-        logger.info("Added signal handlers for cleaning up on exit...")
-    except ValueError:
-        logger.warn("Error adding signal handlers for cleaning up on exit...")
+    signal.signal(signal.SIGINT, signal.SIG_DFL)
 
 
 def create_startup_lock(port):
